+++ conf/config.yaml
@@ -23,11 +23,11 @@ mode: preprocess                  # one of: mask_gen / preprocess / train / infe
 tasks: # TODO: make the other tasks bool key/values
   mask_gen:
     sync: false # skip for now; copies the most up-to-date version of the files needed to run the mask_gen mode from LTS to here
-    detect_weeds: false
-    unet_segment_weeds: false
-    mask_inspect: false
-    mask_refine: false
-    relabel_test: false
+    detect: false
+    segment: false
+    inspect: false
+    refine: false
+    relabel: false
 
   preprocess:
     pad_gridcrop_resize: false
